Django-App/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-rest-knox==4.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.4.0
pip==21.3
psycopg==3.0.1
pycparser==2.20
pytz==2021.3
setuptools==58.1.0
sqlparse==0.4.2
wheel==0.37.0
